refactor: src/client/ を src/ui/ にリネーム
サーバーサイドレンダリングを入れると、これらのコンポーネントはサーバーでも
描画される。「client」という名前は「ブラウザ専用」と読めてしまい、実際に
直前の timeZone 未指定バグはその思い込みから生まれている。同じ間違いを
将来誘発しないよう改名する。

あわせて src/client.tsx と src/client/index.tsx を1ファイルに統合した。
前者は後者を import するだけの中継で、vite の input が src/client.tsx を
指しているために存在していたもの。

ロジックの変更はなく、テストも無修正で通る。
