+++ worker/src/ui/DeviceDetail/index.tsx
