+++ worker/src/ui/components/DeviceStatusCard.tsx
