+++ worker/src/ui/polling.ts
